optimize grbl::send(), and move to separate thread
