+++ Cargo.toml
@@ -11,3 +11,4 @@ serial = "0.4"
 toml = "0.5.8"
 serde = { version = "1.0.120", features = ["derive"] }
 iced = { version = "0.2", features = ["tokio"] }
+chrono = "0.4.19"
